@@ -2,7 +2,13 @@
 
 use Illuminate\Http\Request;
 use Illuminate\Support\Facades\Route;
-use App\User;
+use App\Http\Controllers\Api\AuthController;
+use App\Http\Controllers\Api\PostController;
+use App\Http\Controllers\Api\CategoryController;
+use App\Http\Controllers\Api\TagController;
+use App\Http\Controllers\Api\CommentController;
+
+
 /*
 |--------------------------------------------------------------------------
 | API Routes
@@ -14,38 +20,38 @@ use App\User;
 |
 */
 
-Route::middleware('auth:api')->get('users', function (Request $request) {
+Route::middleware('auth:api')->get('/user', function (Request $request) {
     return $request->user();
 });
 
 Route::group(['middleware' => ['jwt.auth', 'api-header']], function () {
     // all routes to protected resources are registered here
-    Route::post('posts', 'PostController@createPost');
-    Route::get('posts/{id}/edit', 'PostController@editPost');
-    Route::put('posts/{id}', 'PostController@updatePost');
-    Route::delete('posts/{id}', 'PostController@deletePost');
+    Route::post('posts', [PostController::class, 'createPost']);
+    Route::get('posts/{id}/edit', [PostController::class, 'editPost']);
+    Route::put('posts/{id}', [PostController::class, 'updatePost']);
+    Route::delete('posts/{id}', [PostController::class, 'deletePost']);
     // Comment
-    Route::post('comments/{id}', 'CommentController@createComment');
-    Route::post('comments/reply/{id}', 'CommentController@createReplyComment');
+    Route::post('comments/{id}', [CommentController::class, 'createComment']);
+    Route::post('comments/reply/{id}', [CommentController::class, 'createReplyComment']);
 });
 
 Route::group(['middleware' => 'api-header'], function () {
     // The registration and login requests doesn't come with tokens
     // as users at that point have not been authenticated yet
     // Therefore the jwtMiddleware will be exclusive of them
-    Route::post('users/login', 'AuthController@login');
-    Route::post('users/register', 'AuthController@register');
+    Route::post('users/login', [AuthController::class, 'login']);
+    Route::post('users/register', [AuthController::class, 'register']);
     // Posts
-    Route::get('posts', 'PostController@fetchPost');
-    Route::get('posts/{id}', 'PostController@detailPost');
+    Route::get('posts', [PostController::class, 'fetchPost']);
+    Route::get('posts/{id}', [PostController::class, 'detailPost']);
     // Tags
-    Route::get('tags', 'TagController@fetchTag');
-    Route::get('tags/{id}', 'TagController@singleTag');
+    Route::get('tags', [TagController::class, 'fetchTag']);
+    Route::get('tags/{id}', [TagController::class, 'singleTag']);
     // Categories
-    Route::get('categories', 'CategoryController@fetchCategory');
+    Route::get('categories', [CategoryController::class, 'fetchCategory']);
     // Comments
-    Route::get('comments/{id}', 'CommentController@fetchComment');
+    Route::get('comments/{id}', [CommentController::class, 'fetchComment']);
     // Users
-    Route::get('users', 'AuthController@fetchUser');
-    Route::get('users/{id}/{user_name}', 'AuthController@singleUser');
+    Route::get('users', [AuthController::class, 'fetchUser']);
+    Route::get('users/{id}/{user_name}', [AuthController::class, 'singleUser']);
 });
