@@ -1,15 +1,16 @@
 <?php
 
-namespace App\Http\Controllers;
+namespace App\Http\Controllers\Api;
 
+use App\Http\Controllers\Controller;
 use Illuminate\Http\Request;
 use Illuminate\Support\Facades\Hash;
 use Illuminate\Support\Facades\Validator;
 use JWTAuth;
 use JWTAuthException;
-use App\User;
-use App\Role;
-use App\Http\Resources\ProfileResource;
+use App\Models\User;
+use App\Models\Role;
+use App\Http\Resources\UserResource;
 
 class AuthController extends Controller
 {
@@ -118,7 +119,8 @@ class AuthController extends Controller
     public function singleUser($id, $user_name)
     {
         //$user = JWTAuth::toUser($request->token);
-        $user = new ProfileResource(User::where('id', $id)->where('user_name', $user_name)->firstOrFail());
+        $user = new UserResource(User::where('id', $id)->where('user_name', $user_name)->firstOrFail());
         return response()->json(['success' => true, 'data' => $user], 200);
     }
 }
+
