@@ -2,6 +2,7 @@
 
 namespace App\Http\Middleware;
 
+use Illuminate\Http\Request;
 use Closure;
 use JWTAuth;
 use Exception;
@@ -15,22 +16,20 @@ class JWTAuthMiddleware
      * @param  \Closure  $next
      * @return mixed
      */
-    public function handle($request, Closure $next)
+    public function handle(Request $request, Closure $next)
     {
         try {
-            //$user = JWTAuth::toUser($request->input('token'));
-            $user = JWTAuth::toUser($request->token);
-            //$user = JWTAuth::parseToken()->authenticate();
+            $user = JWTAuth::toUser($request->input('token'));
         } catch (Exception $e) {
             if ($e instanceof \Tymon\JWTAuth\Exceptions\TokenInvalidException){
                 return $next($request);
-                return response()->json(['errors' => 'Invalid_Token']);
+                return response()->json(['error'=>'Invalid_Token']);
             }else if ($e instanceof \Tymon\JWTAuth\Exceptions\TokenExpiredException){
                 return $next($request);
-                return response()->json(['errors' => 'Token_Expried']);
+                return response()->json(['error'=>'Token_Expried']);
             }else{
                 return $next($request);
-                return response()->json(['errors' => 'Unknown_Error']);
+                return response()->json(['error'=>'Unknown_Error']);
             }
         }
         return $next($request);
