@@ -5,8 +5,8 @@ namespace App\Http\Resources;
 use Illuminate\Http\Resources\Json\JsonResource;
 use App\Http\Resources\PostTagResource;
 use App\Http\Resources\PostCategoryResource;
-use App\PostCategory;
-use App\Category;
+use App\Models\PostCategory;
+use App\Models\Category;
 
 class PostResource extends JsonResource
 {
