@@ -1,6 +1,6 @@
 <?php
 
-namespace App;
+namespace App\Models;
 
 use Illuminate\Database\Eloquent\Model;
 
@@ -14,11 +14,11 @@ class Category extends Model
 
     public function PostCategory()
     {
-    	return $this->hasMany('App\PostCategory', 'category_id', 'id');
+    	return $this->hasMany('App\Models\PostCategory', 'category_id', 'id');
     }
 
     public function Post()
     {
-        return $this->belongsToMany('App\Post', 'post_category');
+        return $this->belongsToMany('App\Models\Post', 'post_category');
     }
 }
