@@ -1,6 +1,6 @@
 <?php
 
-namespace App;
+namespace App\Models;
 
 use Illuminate\Database\Eloquent\Model;
 
@@ -14,11 +14,11 @@ class RolePermission extends Model
 
     public function Role()
     {
-    	return $this->belongsTo('App\Role', 'role_id', 'id');
+    	return $this->belongsTo('App\Models\Role', 'role_id', 'id');
     }
 
     public function Permission()
     {
-    	return $this->belongsTo('App\Permission', 'permission_id', 'id');
+    	return $this->belongsTo('App\Models\Permission', 'permission_id', 'id');
     }
 }
