@@ -41,12 +41,6 @@ return [
             'provider' => 'users',
         ],
 
-        //'api' => [
-        //    'driver' => 'token',
-        //    'provider' => 'users',
-        //    'hash' => false,
-        //],
-
         'api' => [
             'driver' => 'jwt',
             'provider' => 'users',
@@ -74,7 +68,7 @@ return [
     'providers' => [
         'users' => [
             'driver' => 'eloquent',
-            'model' => App\User::class,
+            'model' => App\Models\User::class,
         ],
 
         // 'users' => [
